requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
bson==0.5.10
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.7
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
Django==5.2.7
django-allauth==65.12.0
django-appconf==1.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-imagekit==6.0.0
django-otp==1.6.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
kombu==5.5.4
packaging==25.0
pilkit==3.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
redis==5.0.1
requests==2.33.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
